Allow extended numeric terms.
Permit arbitrary localized versions of numeric terms, distinguished
by a hyphen and two-digit extension as suffix. This is useful with
the changes in the dev_jurisdiction branch, to enable document
numbers to be customized according to the issuing authority.

This change depends on the checkin:

  f2d7e2297a9b23f9cea97d3d2e2879458f96541a

in the dev_00_terms_refactor branch on github.