Enable gender-based discrimination.
This change enables two forms of gender support. For ordinal
suffixes, multiple forms may be specified with the gender-form=
attribute. The specific form used is determined by the gender
set by gender= attribute, if any, set on the term associated
with the numeric variable (or the month name in the case of a
day within a date) to which the ordinal suffix will be applied.

For ordinary non-numeric variables, the gender-form= variant
of the term is selected according to the gender= attribute,
if any, set for the item type in the locale.