Conditional match of jurisdiction field against a controlled list.
This change introduces a new file to the schema, csl-jurisdiction.rnc,
with content conforming to the proposed URN:LEX specification:

  http://tools.ietf.org/html/draft-spinosa-urn-lex-03

The domain addresses of several major international rule-making
bodies have been added to the list. This permits discrimination
of cite forms according to the issuing jurisdiction, which is
necessary for the correct formatting of legal citations across
national boundaries.