Allow multiple layouts discriminated by locale.
In publishing outside of the English domain (where romanization and
the imposition of native cite conventions is the norm), cites to
foreign works are often presented a the form appropriate to their
language or country of origin. This change permits multiple cs:layout
nodes to be used, each with one or more valid CSL locales on a
locale= attribute. The default layout has no locale= attribute,
and must be included as the last layout specified.